Added Weighted mutations. Currently only BinJAFS is weight awear. Will extend to other stats as it makes sense to do so.

+++ src/at/mabs/external/StatsExport.java
@@ -108,7 +108,7 @@ public class StatsExport {
 				for(StatsCollector sc:collectors){
 					sc.init();
 				}
-				int snps = clm.getSegSiteCount();
+				int snps = (int)clm.getSegSiteCount();
 				Random rand = new Random64();
 				for (int i = 0; i < clm.getRepeats(); i++) {
 					int start = rand.nextInt(mutations.size());
